requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.25
python-dotenv==1.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.2
python-multipart==0.0.6
email-validator==2.1.0.post1
pydantic==2.6.1
httpx==0.27.0
svix==1.19.0
requests==2.31.0
selenium==4.18.1
webdriver-manager==4.0.1
PySide6==6.6.2
